+++ package.json
@@ -13,8 +13,8 @@
     "lint": "eslint . --ext .ts",
     "fix": "eslint --fix . --ext .ts",
     "check": "tsc --noEmit -p tsconfig.worker.json && tsc --noEmit -p tsconfig.node.json",
-    "test": "vitest",
-    "e2e": "vitest --config ./test/e2e/vitest.e2e.config.ts",
+    "test": "vitest --config ./test/vitest.config.ts",
+    "e2e": "vitest --config ./e2e/vitest.config.ts",
     "deploy_commands": "tsx scripts/deploy_commands.ts",
     "generate_test_keys": "tsx scripts/generate_test_keys.ts",
     "publish:test": "wrangler deploy --env test",
