+++ test/vitest.config.ts
@@ -2,14 +2,10 @@ import { defineConfig } from 'vitest/config';
 
 export default defineConfig({
   test: {
-    include: ['test/e2e/**/*.test.ts'],
+    include: ['test/**/*.test.ts'],
     globals: true,
     environment: 'node',
-    setupFiles: ['test/e2e/setup.e2e.ts'],
+    setupFiles: [],
     watch: false,
-    testTimeout: 15000,
-    sequence: {
-      concurrent: false,
-    },
   },
 });
