+++ tsconfig.node.json
@@ -9,7 +9,6 @@
   "include": [
     "scripts/**/*.ts",
     "test/**/*.ts",
-    "vitest.config.ts",
-    "test/e2e/vitest.e2e.config.ts"
+    "e2e/**/*.ts"
   ]
 }
